backend/ltml: Provide single metadata builder
